+++ src/lib.ts
@@ -1,4 +1,1 @@
 export const NO_VALUE: unique symbol = Symbol.for("ORNAMENT_NO_VALUE");
-export const EVENT_BUS_TARGET: unique symbol = Symbol.for("ORNAMENT_BUS");
-export const isArray = Array.isArray;
-export const EMPTY_OBJ = {};